feat(ui): share ActivityMenu across PRs and Issues, add dashboard refresh
Move the activity dropdown into a shared component so it now renders on
both the PRs and Issues pages. Fold the header poll-interval readout into
the dropdown to remove header clutter, split AppHeader into primary
(PRs/Issues/Releases) and secondary (Logs/Settings) nav, keep the PR-list
tab labels on a single line, and add a refresh button to the dashboard
header that invalidates the dashboard query keys.
