+++ client/src/pages/dashboard.tsx
@@ -2,17 +2,13 @@ import { memo, useEffect, useMemo, useRef, useState, type MouseEvent } from "rea
 import { Link } from "wouter";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import * as Collapsible from "@radix-ui/react-collapsible";
-import { Activity as ActivityIcon, AlertTriangle, Bot, ChevronDown, ChevronUp, Trash2 } from "lucide-react";
+import { AlertTriangle, Bot, ChevronDown, ChevronUp, Loader2, RefreshCw, Trash2 } from "lucide-react";
 import { queryClient, apiRequest, fetchJson } from "@/lib/queryClient";
 import type { ActivityItem, ActivitySnapshot, Config, FeedbackItem, HealingSession, Issue, LogEntry, OperatorWarning, PR, PRQuestion, RuntimeState, WatchedRepo } from "@shared/schema";
 import { AppHeader } from "@/components/AppHeader";
 import { OnboardingPanel } from "@/components/OnboardingPanel";
 import { UpdateBanner } from "@/components/UpdateBanner";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
+import { ActivityMenu, EMPTY_ACTIVITY_SNAPSHOT } from "@/components/ActivityMenu";
 import { Tooltip, TooltipTrigger, TooltipContent } from "@/components/ui/tooltip";
 import { toast } from "@/hooks/use-toast";
 import {
@@ -86,11 +82,6 @@ function formatStatusLabel(status: PR["status"]): string {
   return "watching";
 }
 
-function formatPollInterval(pollIntervalMs?: number): string {
-  const seconds = Math.max(1, Math.round((pollIntervalMs ?? 120000) / 1000));
-  return `${seconds}s`;
-}
-
 function latestActivityForTarget(activities: ActivityItem[], targetId: string): ActivityItem | undefined {
   return activities.reduce((latest, activity) => {
     if (activity.targetId !== targetId) {
@@ -130,20 +121,12 @@ function getPRFeedbackFailureReason(pr: PR): string | null {
   return failedItem?.statusReason ?? null;
 }
 
-const EMPTY_ACTIVITY_SNAPSHOT: ActivitySnapshot = {
-  failed: [],
-  inProgress: [],
-  queued: [],
-  warnings: [],
-  generatedAt: "",
-};
 const MAX_VISIBLE_LOGS = 200;
 const DRAIN_PAUSED_LABEL = "Paused";
 const DRAIN_PAUSED_TITLE = "Paused by drain mode";
 const MANUAL_RUNS_BLOCKED_COPY = "Manual runs are blocked while global automation is paused.";
 const GLOBAL_DRAIN_PR_COPY = "Background and manual runs are paused by drain mode.";
 const ASK_DRAIN_COPY = "Ask Agent is paused by drain mode.";
-const QUEUED_DRAIN_COPY = "Queued automation is paused until drain mode is disabled.";
 
 const HEALING_TONE_CLASSES: Record<"neutral" | "info" | "warning" | "success" | "danger", string> = {
   neutral: "border-border text-muted-foreground",
@@ -177,171 +160,6 @@ function WatchPausedIndicator() {
     <span className="border border-border px-1.5 py-0 text-[10px] uppercase tracking-wider text-muted-foreground">
       watch paused
     </span>
-  );
-}
-
-function ActivityRow({ activity, queueStatus }: { activity: ActivityItem; queueStatus: QueueStatusView | null }) {
-  const timeLabel = activity.status === "failed"
-    ? formatClock(activity.updatedAt)
-    : activity.status === "in_progress"
-    ? formatClock(activity.startedAt) ?? formatClock(activity.updatedAt)
-    : formatClock(activity.availableAt) ?? formatClock(activity.queuedAt);
-  const content = (
-    <div className="flex min-w-0 items-start gap-2 px-2 py-1.5 text-left">
-      <span
-        className={`mt-1.5 h-1.5 w-1.5 shrink-0 ${
-          activity.status === "failed"
-            ? "bg-destructive"
-            : activity.status === "in_progress"
-              ? "animate-pulse bg-foreground"
-              : "bg-muted-foreground"
-        }`}
-      />
-      <span className="min-w-0 flex-1">
-        <span className="block truncate text-[12px] leading-4 text-foreground">{activity.label}</span>
-        {activity.detail && (
-          <span className="block truncate text-[11px] leading-4 text-muted-foreground">{activity.detail}</span>
-        )}
-        <div className="mt-1">
-          <QueueStatusBadge status={queueStatus} />
-        </div>
-        {activity.status === "failed" && activity.lastError && (
-          <span
-            className="block whitespace-pre-wrap break-words text-[11px] leading-4 text-destructive"
-            title={activity.lastError}
-          >
-            {activity.lastError}
-          </span>
-        )}
-      </span>
-      {timeLabel && (
-        <span className="shrink-0 text-[10px] leading-4 text-muted-foreground">{timeLabel}</span>
-      )}
-    </div>
-  );
-
-  if (activity.targetUrl) {
-    return (
-      <a
-        href={activity.targetUrl}
-        target="_blank"
-        rel="noopener noreferrer"
-        className="block outline-none hover:bg-muted focus:bg-muted focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-1 focus-visible:ring-offset-background"
-      >
-        {content}
-      </a>
-    );
-  }
-
-  return <div>{content}</div>;
-}
-
-function ActivitySection({
-  title,
-  items,
-  emptyLabel,
-  queueStatusById,
-}: {
-  title: string;
-  items: ActivityItem[];
-  emptyLabel: string;
-  queueStatusById: Map<string, QueueStatusView>;
-}) {
-  return (
-    <div className="py-1">
-      <div className="px-2 py-1 text-[10px] uppercase tracking-wider text-muted-foreground">{title}</div>
-      {items.length > 0 ? (
-        <div className="max-h-52 overflow-y-auto">
-          {items.map((activity) => (
-            <ActivityRow key={activity.id} activity={activity} queueStatus={queueStatusById.get(activity.targetId) ?? null} />
-          ))}
-        </div>
-      ) : (
-        <div className="px-2 pb-2 text-[11px] text-muted-foreground">{emptyLabel}</div>
-      )}
-    </div>
-  );
-}
-
-function ActivityMenu({
-  activities,
-  onClearFailed,
-  isClearingFailed,
-  globalDrainMode,
-  queueStatusById,
-}: {
-  activities: ActivitySnapshot;
-  onClearFailed: () => void;
-  isClearingFailed: boolean;
-  globalDrainMode: boolean;
-  queueStatusById: Map<string, QueueStatusView>;
-}) {
-  const failedCount = activities.failed.length;
-  const inProgressCount = activities.inProgress.length;
-  const queuedCount = activities.queued.length;
-  const totalCount = failedCount + inProgressCount + queuedCount;
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger
-        className="inline-flex items-center gap-1 border border-border px-2 py-0.5 text-[11px] text-muted-foreground hover:text-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-1 focus-visible:ring-offset-background"
-        aria-label="Open activity menu"
-        data-testid="activity-menu-trigger"
-      >
-        <ActivityIcon className="h-3 w-3" aria-hidden="true" />
-        <span>activity</span>
-        <span className="text-foreground">{totalCount}</span>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end" className="w-80 p-0">
-        <div className="border-b border-border px-2 py-2">
-          <div className="flex items-center justify-between gap-2">
-            <div className="text-[12px] font-medium">Activities</div>
-            {failedCount > 0 && (
-              <button
-                type="button"
-                onClick={onClearFailed}
-                disabled={isClearingFailed}
-                className="border border-border px-1.5 py-0.5 text-[10px] uppercase tracking-wider text-muted-foreground hover:text-foreground disabled:cursor-not-allowed disabled:opacity-50 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
-                data-testid="clear-failed-activities"
-              >
-                {isClearingFailed ? "clearing" : "clear failed"}
-              </button>
-            )}
-          </div>
-          <div className="text-[11px] text-muted-foreground">
-            {failedCount} failed / {inProgressCount} in progress / {queuedCount} queued
-          </div>
-        </div>
-        <ActivitySection
-          title="Failed"
-          items={activities.failed}
-          emptyLabel="No failed activities."
-          queueStatusById={queueStatusById}
-        />
-        <div className="border-t border-border" />
-        <ActivitySection
-          title="In progress"
-          items={activities.inProgress}
-          emptyLabel="No activities running right now."
-          queueStatusById={queueStatusById}
-        />
-        <div className="border-t border-border" />
-        <ActivitySection
-          title="Queued"
-          items={activities.queued}
-          emptyLabel="Queue is empty."
-          queueStatusById={queueStatusById}
-        />
-        {globalDrainMode && queuedCount > 0 && (
-          <div
-            className="border-t border-border px-2 py-2 text-[11px] text-muted-foreground"
-            data-testid="activity-drain-note"
-          >
-            {QUEUED_DRAIN_COPY}
-          </div>
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
   );
 }
 
@@ -1287,6 +1105,24 @@ export default function Dashboard() {
   const [viewMode, setViewMode] = useState<"active" | "issues" | "archived">("active");
   const [prNumberSearch, setPrNumberSearch] = useState("");
   const [areErrorsRolledUp, setAreErrorsRolledUp] = useState(false);
+  const [isRefreshing, setIsRefreshing] = useState(false);
+
+  const handleRefreshDashboard = async () => {
+    setIsRefreshing(true);
+    try {
+      await Promise.all([
+        queryClient.invalidateQueries({ queryKey: ["/api/prs"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/prs/archived"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/issues"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/activities"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/healing-sessions"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/runtime"] }),
+        queryClient.invalidateQueries({ queryKey: ["/api/repos/settings"] }),
+      ]);
+    } finally {
+      setIsRefreshing(false);
+    }
+  };
 
   const { data: prs = [], isLoading } = useQuery<PR[]>({
     queryKey: ["/api/prs"],
@@ -1454,14 +1290,9 @@ export default function Dashboard() {
       <AppHeader
         active="prs"
         status={(
-          <>
-            <span>
-              <span className="font-mono text-foreground">{prs.length}</span> PR{prs.length !== 1 ? "s" : ""} / <span className="font-mono text-foreground">{repos.length}</span> repo{repos.length !== 1 ? "s" : ""}
-            </span>
-            <span>
-              poll <span className="font-mono text-foreground/80">{formatPollInterval(config?.pollIntervalMs)}</span>
-            </span>
-          </>
+          <span>
+            <span className="font-mono text-foreground">{prs.length}</span> PR{prs.length !== 1 ? "s" : ""} / <span className="font-mono text-foreground">{repos.length}</span> repo{repos.length !== 1 ? "s" : ""}
+          </span>
         )}
         actions={(
           <>
@@ -1497,12 +1328,23 @@ export default function Dashboard() {
                 <span className="font-mono">{activeErrorCount}</span>
               </button>
             )}
+            <button
+              type="button"
+              onClick={() => { void handleRefreshDashboard(); }}
+              disabled={isRefreshing}
+              data-testid="button-refresh-dashboard"
+              className="inline-flex items-center gap-1 rounded-md border border-border px-2 py-0.5 text-[11px] uppercase tracking-wider text-muted-foreground hover:text-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-offset-1 focus-visible:ring-offset-background disabled:opacity-50"
+            >
+              {isRefreshing ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <RefreshCw className="h-3.5 w-3.5" />}
+              refresh
+            </button>
             <ActivityMenu
               activities={activities}
               onClearFailed={() => clearFailedActivitiesMutation.mutate()}
               isClearingFailed={clearFailedActivitiesMutation.isPending}
               globalDrainMode={globalDrainMode}
               queueStatusById={queueStatusById}
+              pollIntervalMs={config?.pollIntervalMs}
             />
           </>
         )}
@@ -1528,7 +1370,7 @@ export default function Dashboard() {
               type="button"
               onClick={() => setViewMode("active")}
               data-testid="tab-active"
-              className={`flex-1 px-3 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
+              className={`flex-1 whitespace-nowrap px-2 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
                 viewMode === "active"
                   ? "bg-muted text-foreground shadow-[inset_0_-2px_0_0_hsl(var(--primary))]"
                   : "text-muted-foreground hover:text-foreground"
@@ -1540,19 +1382,19 @@ export default function Dashboard() {
               type="button"
               onClick={() => setViewMode("issues")}
               data-testid="tab-on-issues"
-              className={`flex-1 px-3 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
+              className={`flex-1 whitespace-nowrap px-2 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
                 viewMode === "issues"
                   ? "bg-muted text-foreground shadow-[inset_0_-2px_0_0_hsl(var(--primary))]"
                   : "text-muted-foreground hover:text-foreground"
               }`}
             >
-              On Issues ({issueLinkedPRs.length})
+              Issues ({issueLinkedPRs.length})
             </button>
             <button
               type="button"
               onClick={() => setViewMode("archived")}
               data-testid="tab-archived"
-              className={`flex-1 px-3 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
+              className={`flex-1 whitespace-nowrap px-2 py-2 text-[11px] uppercase tracking-wider transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring focus-visible:ring-inset ${
                 viewMode === "archived"
                   ? "bg-muted text-foreground shadow-[inset_0_-2px_0_0_hsl(var(--primary))]"
                   : "text-muted-foreground hover:text-foreground"
